Move nab_index cache tests to python workspace
